+++ openquake/hazardlib/source/base.py
@@ -21,6 +21,8 @@ import abc
 import math
 import numpy
 from openquake.baselib.slots import with_slots
+from openquake.hazardlib.geo import Point
+from openquake.hazardlib.source.rupture import ParametricProbabilisticRupture
 
 
 @with_slots
@@ -109,22 +111,13 @@ class BaseSeismicSource(metaclass=abc.ABCMeta):
         :param ir_monitor: a monitor object for .iter_ruptures()
         :yields: pairs (rupture, num_occurrences[num_samples])
         """
-        mutex_weight = getattr(self, 'mutex_weight', 1)
-        with ir_monitor:
-            ruptures = list(self.iter_ruptures())
         tom = getattr(self, 'temporal_occurrence_model', None)
         serials = numpy.arange(self.serial, self.serial + self.num_ruptures)
-
         if tom:  # time-independent source
-            rates = numpy.array([rup.occurrence_rate for rup in ruptures])
-            numpy.random.seed(self.serial)
-            occurs = numpy.random.poisson(rates * tom.time_span * eff_num_ses)
-            for rup, serial, num_occ in zip(ruptures, serials, occurs):
-                if num_occ:
-                    rup.serial = serial  # used as seed
-                    yield rup, num_occ
+            yield from self.sample_ruptures_poissonian(serials, eff_num_ses)
         else:  # time-dependent source
-            for rup, serial in zip(ruptures, serials):
+            mutex_weight = getattr(self, 'mutex_weight', 1)
+            for rup, serial in zip(self.iter_ruptures(), serials):
                 numpy.random.seed(serial)
                 occurs = rup.sample_number_of_occurrences(eff_num_ses)
                 if mutex_weight < 1:
@@ -134,6 +127,49 @@ class BaseSeismicSource(metaclass=abc.ABCMeta):
                 if num_occ:
                     rup.serial = serial  # used as seed
                     yield rup, num_occ
+
+    def sample_ruptures_poissonian(self, serials, eff_num_ses):
+        """
+        :param eff_num_ses: number of stochastic event sets * number of samples
+        :yields: pairs (rupture, num_occurrences[num_samples])
+        """
+        tom = self.temporal_occurrence_model
+        if not hasattr(self, 'nodal_plane_distribution'):  # fault
+            ruptures = list(self.iter_ruptures())
+            rates = numpy.array([rup.occurrence_rate for rup in ruptures])
+            numpy.random.seed(self.serial)
+            occurs = numpy.random.poisson(rates * tom.time_span * eff_num_ses)
+            for rup, serial, num_occ in zip(ruptures, serials, occurs):
+                if num_occ:
+                    rup.serial = serial  # used as seed
+                    yield rup, num_occ
+            return
+        # else (multi)point sources and area sources
+        rup_args = []
+        rates = []
+        for src in self:
+            for mag, mag_occ_rate in src.get_annual_occurrence_rates():
+                for np_prob, np in src.nodal_plane_distribution.data:
+                    for hc_prob, hc_depth in src.hypocenter_distribution.data:
+                        args = (mag_occ_rate, np_prob, hc_prob,
+                                mag, np, hc_depth, src)
+                        rup_args.append(args)
+                        rates.append(mag_occ_rate * np_prob * hc_prob)
+        eff_rates = numpy.array(rates) * tom.time_span * eff_num_ses
+        numpy.random.seed(self.serial)
+        occurs = numpy.random.poisson(eff_rates)
+        for num_occ, args, rate, ser in zip(occurs, rup_args, rates, serials):
+            if num_occ:
+                mag_occ_rate, np_prob, hc_prob, mag, np, hc_depth, src = args
+                hc = Point(latitude=src.location.latitude,
+                           longitude=src.location.longitude,
+                           depth=hc_depth)
+                surface = src._get_rupture_surface(mag, np, hc)
+                rup = ParametricProbabilisticRupture(
+                    mag, np.rake, src.tectonic_region_type, hc,
+                    surface, rate, tom)
+                rup.serial = ser  # used as seed
+                yield rup, num_occ
 
     @abc.abstractmethod
     def get_one_rupture(self, rupture_mutex=False):
